key ordering fix

+++ cmd/tools/book-items/main.go
@@ -8,12 +8,14 @@ import (
 	"cloud.google.com/go/firestore"
 	"github.com/easterthebunny/spew-order/internal/persist"
 	"github.com/easterthebunny/spew-order/internal/persist/firebase"
+	"github.com/easterthebunny/spew-order/pkg/types"
 )
 
 var (
-	account   = flag.String("account", "spew", "Environment naming prefix.")
 	projectID = flag.String("project", "", "Google project id.")
+	account   = flag.String("account", "", "account uuid")
 	delete    = flag.Bool("delete", false, "delete book items")
+	print     = flag.Bool("print", false, "print order book")
 )
 
 func main() {
@@ -26,38 +28,72 @@ func main() {
 	}
 
 	arepo := firebase.NewAccountRepository(client)
-	orepo := arepo.Orders(&persist.Account{ID: *account})
 	brepo := firebase.NewBookRepository(client)
 
-	orders, err := orepo.GetOrdersByStatus(ctx, persist.StatusOpen, persist.StatusPartial, persist.StatusCanceled)
-	if err != nil {
-		panic(err)
+	if *print {
+
+		/*
+		   BUY
+		   0.0056 ------> 60
+		   0.0055 --->    32
+
+		   SELL
+		*/
+		item := persist.BookItem{
+			Order: types.Order{
+				OrderRequest: types.OrderRequest{
+					Base:   types.SymbolBitcoin,
+					Target: types.SymbolEthereum,
+				},
+			},
+			ActionType: types.ActionTypeBuy,
+		}
+
+		fmt.Printf("%s\n", item.ActionType.String())
+		head, _ := brepo.GetHeadBatch(ctx, &item, 50)
+		for _, h := range head {
+			switch j := h.Order.Type.(type) {
+			case *types.MarketOrderType:
+				fmt.Printf("MARKET:%s %30s\n", j.Base, j.Quantity)
+			case *types.LimitOrderType:
+				fmt.Printf("%-10s %30s\n", j.Price, j.Quantity)
+			}
+		}
 	}
 
-	for _, order := range orders {
-		bitem := persist.NewBookItem(order.Base)
-		result, err := brepo.BookItemExists(ctx, &bitem)
+	if *account != "" {
+		orepo := arepo.Orders(&persist.Account{ID: *account})
+
+		orders, err := orepo.GetOrdersByStatus(ctx, persist.StatusOpen, persist.StatusPartial, persist.StatusCanceled)
 		if err != nil {
 			panic(err)
 		}
 
-		fmt.Printf("action %s; status: %s; exists on book: %v\n", order.Base.Action, order.Status, result)
-		if order.Status == persist.StatusCanceled && result && *delete {
-			err = brepo.DeleteBookItem(ctx, &bitem)
-			if err != nil {
-				panic(err)
-			}
-		}
-
-		if order.Status == persist.StatusOpen || order.Status == persist.StatusPartial {
-			items, err := brepo.GetHeadBatch(ctx, &bitem, 10)
+		for _, order := range orders {
+			bitem := persist.NewBookItem(order.Base)
+			result, err := brepo.BookItemExists(ctx, &bitem)
 			if err != nil {
 				panic(err)
 			}
 
-			fmt.Printf("%d items found\n", len(items))
-			for _, item := range items {
-				fmt.Printf("%v\n", item)
+			fmt.Printf("action %s; status: %s; exists on book: %v\n", order.Base.Action, order.Status, result)
+			if order.Status == persist.StatusCanceled && result && *delete {
+				err = brepo.DeleteBookItem(ctx, &bitem)
+				if err != nil {
+					panic(err)
+				}
+			}
+
+			if order.Status == persist.StatusOpen || order.Status == persist.StatusPartial {
+				items, err := brepo.GetHeadBatch(ctx, &bitem, 10)
+				if err != nil {
+					panic(err)
+				}
+
+				fmt.Printf("%d items found\n", len(items))
+				for _, item := range items {
+					fmt.Printf("%v\n", item)
+				}
 			}
 		}
 	}
